+++ js/components/template/kanban/AddUser.js
@@ -1,25 +1,10 @@
 import CreateElement from '../../../lib/CreateElement.js';
-import { setter } from '../../../store/team.js';
-import { validateName } from '../../../lib/validators.js';
-import STRINGS from '../../../constant/STRINGS.js';
-import { createTeamMember } from '../../../endpoint/team/service.js';
+import { addMemberHandler } from '../../../eventHandler.js';
 
 const AddUser = ({ teamId }) => {
   const dom = CreateElement('li', { className: 'add-user-button-container' });
 
-  const addMemberHandler = async({ target }) => {
-    if (!target.closest('#add-user-button')) return;
-
-    const name = prompt(STRINGS.memberNamePromptMessage);
-    if (name === null) return;
-
-    const isValid = await validateName(name, addMemberHandler);
-    if (!isValid) return;
-    const newTeamMembers = await createTeamMember({ teamId, name });
-    setter.teamInfo(newTeamMembers);
-  };
-
-  dom.addEventListener('click', addMemberHandler);
+  dom.addEventListener('click', (event) => addMemberHandler(teamId, event));
 
   const render = () => {
     dom.innerHTML = `
