+++ js/components/template/team/AddTeam.js
@@ -1,23 +1,9 @@
-import { postTeam } from '../../../endpoint/team/controller.js';
-import { validateName } from '../../../lib/validators.js';
-import { dispatch } from '../../../store/team.js';
 import CreateElement from '../../../lib/CreateElement.js';
-import STRINGS from '../../../constant/STRINGS.js';
+import { addTeamHandler } from '../../../eventHandler.js';
 
 const AddTeam = (props) => {
 
   const dom = CreateElement('li', { className: 'add-team-button-container' });
-
-  const addTeamHandler = async () => {
-      const name = prompt(STRINGS.teamNamePromptMessage);
-      if (name === null) return;
-
-      const isValid = await validateName(name, addTeamHandler);
-      if (isValid) {
-        await postTeam({ name });
-        dispatch.teamList();
-      }
-  };
 
   dom.addEventListener('click', addTeamHandler);
